fix: cast interval_ms to int to support float interval_minutes

+++ timer.py
@@ -6,7 +6,7 @@ class RestTimer(QObject):
 
     def __init__(self, interval_minutes: int = 20):
         super().__init__()
-        self._interval_ms = interval_minutes * 60 * 1000
+        self._interval_ms = int(interval_minutes * 60 * 1000)
         self._paused = False
         self._timer = QTimer()
         self._timer.timeout.connect(self.rest_due)
@@ -25,7 +25,7 @@ class RestTimer(QObject):
             self._timer.start(self._interval_ms)
 
     def reset(self, interval_minutes: int) -> None:
-        self._interval_ms = interval_minutes * 60 * 1000
+        self._interval_ms = int(interval_minutes * 60 * 1000)
         self._timer.stop()
         self._paused = False
         self._timer.start(self._interval_ms)
